pyproject.toml
[tool.poetry]
name = "dms-cs-lectures"
version = "0.1.0"
description = ""
authors = ["aleacarde"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
python-keycloak = "^4.7.1"
passlib = "^1.7.4"
cachetools = "^5.5.0"
aiohttp = "^3.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
